Command Documentation

Index

BeaData.Bea

BeaData.BeaNipaTable

BeaData.get_nipa_table

BeaData.nipa_metadata_tex

BeaData.table_metadata_tex

BeaData.Bea — Type.

A connection to the U.S. Bureau of Economic Analysis (BEA) Data API.

Constructors

Bea()

Bea(key::AbstractString)

Arguments

key
: Registration key provided by the BEA.

A valid registration key is required to retrieve data from the BEA's API. A key can be obtained by registering at the BEA website.

A default API key can be specified in a ~/.beadatarc file.

Methods

api_url(b::Bls)
: Get the base URL used to connect to the server

api_key(b::Bls)
: Get the API key

BeaData.BeaNipaTable — Type.

A NIPA table with data and metadata returned from a get_nipa_table call.

Fields

tablenum::AbstractString
: NIPA table number

tableid::Int
: API TableID

tabledesc::AbstractString
: The table title (e.g., "Real Gross Domestic Product, Chained Dollars" for Table 1.1.6)

linedesc::OrderedDict
: Dictionary of descriptions for each line of the table

tablenotes::Any
: Table notes, if any

frequency::AbstractString
: "A" or "Q"

startyear::Int

endyear::Int

df::DataFrame
: the data values from the table; column names are the line numbers from the table, the first column contains the date for each observation in Julia
Date
format

BeaData.get_nipa_table — Method.

get_nipa_table(b::Bea, TableID::Int, frequency::AbstractString,
    startyear::Int, endyear::Int)

Request a NIPA table from the BEA data API.

Arguments

b
: a BEA API connection

TableID
: the integer Table ID for the desired NIPA table

frequency
: "A" for annual, "Q" for quarerly

startyear
: first year of data requested, in YYYY format

endyear
: last year of data requested, in YYYY format

Returns

An object of type
BeaNipaTable

BeaData.nipa_metadata_tex — Method.

nipa_metadata_tex(b::Bea)

Arguments

b
: a BEA API connection

Returns

A .tex file with the parmater list for the NIPA dataset and parameter values for the TableID parameter. The file is written to the curent working directory.

BeaData.table_metadata_tex — Method.

table_metadata_tex(bnt::BeaNipaTable)

Arguments

bnt
: a BEA NIPA Table object

Returns

A .tex file with metadata (table name and description, line numbers and descriptions, and table notes) for a NIPA table. The file is written to the curent working directory.