Index

Command Documentation

Index

BeaData.Bea

BeaData.BeaTable

BeaData.get_bea_datasets

BeaData.get_bea_parameterlist

BeaData.get_nipa_table

BeaData.Bea — Type.

type Bea

A connection to the U.S. Bureau of Economic Analysis (BEA) Data API.

Constructors

Bea()

Bea(key::AbstractString)

Arguments

key: Registration key provided by the BEA.

A valid registration key is required to retrieve data from the BEA's API. A key can be obtained by registering at the BEA website.

A default API key can be specified in a ~/.beadatarc file.

Fields

url

key

source

BeaData.BeaTable — Type.

type BeaTable

A NIPA table with data and metadata returned from a get_nipa_table call.

Fields

tablenum - NIPA table number

tablename - API TableName

tabledesc - The table title (e.g., 'Real Gross Domestic Product, Chained Dollars' for Table 1.1.6)

linedesc - OrderedDict of descriptions for each line of the table

tablenotes - Table notes, if any

frequency

startyear

endyear

df - DataFrame containing the data values from the table; column names are the line numbers from the table, the first column contains the date for each observation in Julia Date format

source

BeaData.get_bea_datasets — Method.

get_bea_datasets(b::Bea)

Return a Dict of dataset names and descriptions.

Arguments

b – a Bea connection

source

BeaData.get_bea_parameterlist — Method.

get_bea_parameterlist(b::BeaData.Bea, dataset::String)

Return a Dict of parameters for dataset.

Arguments

b – a Bea connection

dataset – String indicating the dataset ID.

source

BeaData.get_nipa_table — Method.

get_nipa_table(b::Bea, TableName::AbstractString, frequency::AbstractString, startyear::Int, endyear::Int)

Request a NIPA table from the BEA data API and return an object of type BeaNipaTable.

Arguments

b – a Bea connection

TableName – the Table Name (string) for the desired NIPA table

frequency – "A" for annual, "Q" for quarerly

startyear – first year of data requested, in YYYY format

endyear – last year of data requested, in YYYY format

source