Command Documentation

Index

BeaData.Bea

BeaData.BeaNipaTable

BeaData.get_bea_datasets

BeaData.get_bea_parameterlist

BeaData.get_nipa_table

BeaData.nipa_metadata_tex

BeaData.parse_data_dict

BeaData.table_metadata_tex

BeaData.Bea — Type.

type Bea

A connection to the U.S. Bureau of Economic Analysis (BEA) Data API.

Constructors

Bea()

Bea(key::AbstractString)

Arguments

key: Registration key provided by the BEA.

A valid registration key is required to retrieve data from the BEA's API. A key can be obtained by registering at the BEA website.

A default API key can be specified in a ~/.beadatarc file.

Fields

url

key

dataset

source

BeaData.BeaNipaTable — Type.

type BeaNipaTable

A NIPA table with data and metadata returned from a get_nipa_table call.

Fields

tablenum - NIPA table number

tableid - API TableID

tabledesc - The table title (e.g., 'Real Gross Domestic Product, Chained Dollars' for Table 1.1.6)

linedesc - OrderedDict of descriptions for each line of the table

tablenotes - Table notes, if any

frequency

startyear

endyear

df - DataFrame containing the data values from the table; column names are the line numbers from the table, the first column contains the date for each observation in Julia Date format

source

BeaData.get_bea_datasets — Method.

get_bea_datasets(b::Bea)

Return, in a DataFrame, a list of IDs and descriptions for datasets accessible from the BEA data API.

Arguments

b – a Bea connection

source

BeaData.get_bea_parameterlist — Method.

get_bea_parameterlist(b::BeaData.Bea, dataset::String)

Return, in a DataFrame, a list of parameters for dataset.

Arguments

b – a Bea connection

dataset – String indicating the dataset ID.

Returns

A DataFrame listing the following:

Parameter ID

Parmameter description

required – 1 for yes, 0 for no

default_value – empty if none

source

BeaData.get_nipa_table — Method.

get_nipa_table(b::Bea, TableID::Int, frequency::AbstractString, startyear::Int, endyear::Int)

Request a NIPA table from the BEA data API and return an object of type BeaNipaTable.

Arguments

b – a Bea connection

TableID – the integer Table ID for the desired NIPA table

frequency – "A" for annual, "Q" for quarerly

startyear – first year of data requested, in YYYY format

endyear – last year of data requested, in YYYY format

source

BeaData.nipa_metadata_tex — Method.

nipa_metadata_tex(b::Bea)

Write, to the current working directory, a .tex file with the parmater list for the NIPA dataset and parameter values for the TableID parameter.

Arguments

b – a Bea connection

source

BeaData.table_metadata_tex — Method.

table_metadata_tex(bnt::BeaNipaTable)

Write, to the current working directory, a .tex file with metadata (table name and description, line numbers and descriptions, and table notes) for the NIPA table contained in bnt.

Arguments

bnt – a BeaNipaTable object

source

BeaData.parse_data_dict — Method.

parse_data_dict(dict::Dict)

Extract information for a single observation and return as a tuple. (Internal method for get_nipa_table.)

source